+++ python/GafferSceneUI/RenderPassEditor.py
@@ -452,7 +452,10 @@ class RenderPassEditor( GafferUI.NodeSetEditor ) :
 		else :
 			renderPassPlug = script["variables"]["renderPass"]
 
-		renderPassPlug["value"].setValue( selectedPassNames[0] )
+		with Gaffer.Context( self.getContext() ) :
+			currentRenderPass = renderPassPlug["value"].getValue()
+
+		renderPassPlug["value"].setValue( selectedPassNames[0] if selectedPassNames[0] != currentRenderPass else "" )
 
 	## \todo Consider consolidating this with `LightEditor.__editSelectedCells()`.
 	# The main difference being the name of the context variable that is set before inspection,