+++ Changes.md
@@ -18,6 +18,7 @@ Improvements
     - The `data_seed`, `proc_seed`, `obj_seed`, and `face_seed` parameters of the `color_jitter` shader.
 - EditScope : Added summaries of set membership edits in the NodeEditor.
 - LightEditor : Mute and solo columns now accurately reflect the presence of the `light:mute` attribute (for the Mute column) and membership in the `soloLights` set (for the Solo column) for all scene locations, not just for lights.
+- RenderPassEditor : The currently active render pass can now be unset by double clicking on its green dot in the "Active" column.
 
 Fixes
 -----